# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AGOGO API",
    "description": "API for AGOGO's APP."
  },
  "host": "agogo-app-7648993f1c5c.herokuapp.com/",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/dietitian/login": {
      "get": {
        "tags": ["Dietitian API"],
        "summary": "Login for Dietitian",
        "description": "Authenticate and retrieve details for a dietitian based on email and password.",
        "parameters": [
          {
            "name": "dietitian_ID",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the dietitian"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Dietitian"
            }
          },
          "400": {
            "$ref": "#/definitions/apiResponse"
          }
        }
      }
    },
    "/dietitian/getDietitians": {
      "get": {
        "tags": ["Dietitian API"],
        "summary": "Get List of system Dietitians",
        "description": "Get List of system Dietitians.",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "Type":"array",
              "items":{
              "$ref": "#/definitions/Dietitian"
              }
            }
          },
          "400": {
            "description": "Error."
          }
        }
      }
    },
    "/dietitian/patients": {
      "get": {
        "tags": ["Dietitian API"],
        "summary": "Fetch dietitian's patients",
        "description": "Fetch all patients of a specific dietitian",
        "parameters": [
          {
            "name": "dietitian_ID",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the dietitian"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
             "schema": {
                "Type": "array",
                "items":{ 
                "$ref": "#/definitions/Patient"
                }
          }
        }
      }
     }
    },
    "/dietitian/createPatient": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Create a patient",
        "description": "Create a new patient for a specific dietitian",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
               "$ref": "#/definitions/CreatePatient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
             "schema": {
               "$ref": "#/definitions/Patient"
             }
          }
        }
      }
    },
    "/dietitian/deactivatePatient": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Deactivate a patient",
        "description": "Deactivate a specific patient",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "patient_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
                "patient_ID": {
                  "type": "String"
                }
              }
          }
        }
      }
    },
    "/dietitian/activatePatient": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Activate a patient",
        "description": "Activate a specific patient",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "patient_ID": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
             "schema": {
                 "patient_ID": {
                   "type": "String"
                 }
               }
          }
        }
      }
    },
    "/dietitian/addDietitian": {
      "post": {
        "tags": ["Dietitian API"],
        "summary": "Create a Dietitian",
        "description": "Create a new Dietitian",
        "parameters": [
          {
            "name": "dietitianDetails",
            "in": "body",
            "required": true,
            "schema": {
               "$ref": "#/definitions/DietitianCreation"
            }
          }
        ],
        "responses": {
          "200": {
            "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
          }
        }
      }
    },
    "/dietitian/updateDietitian": {
      "put": {
        "tags": ["Dietitian API"],
        "summary": "Update a Dietitian",
        "description": "Update a Dietitian",
        "parameters": [
          {
            "name": "patientDetails",
            "in": "body",
            "required": true,
            "schema": {
               "$ref": "#/definitions/DietitianUPD"
            }
          }
        ],
        "responses": {
          "200": {
            "type": "object",
              "properties": {
                "dietitian_id": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
          }
        }
      }
    },
    "/dietitian/deleteDietitian": {
      "delete": {
        "tags": ["Dietitian API"],
        "summary": "Update a Dietitian",
        "description": "Update a Dietitian",
        "parameters": [
          {
            "name": "Dietitian Id",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_id": {
                  "type": "string"
                }
              }
            }
          }
          
        ],
        "responses": {
          "200": {
            "description": "dietitian with ID: {dietitian_id} deleted successfully."
          }
        }
      }
    },
    "/MealPrep/generateMealPlan": {
      "post": {
        "tags": ["Meal Plan Generation API"],
        "summary": "Generate meal plan using LSM",
        "description": "Generates a meal plan based on protein, carbs, and fat goals using Least Squares Method (LSM).",
        "parameters": [
          {
            "name": "mealPlanDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_ID": {
                  "type": "string"
                },
                "protein_goal": {
                  "type": "number"
                },
                "carbs_goal": {
                  "type": "number"
                },
                "fat_goal": {
                  "type": "number"
                },
                "nbr_days": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/GenMealPrep"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/MealPrep/generateMealPlanFixedLunch": {
      "post": {
        "tags": ["Meal Plan Generation API"],
        "summary": "Generate meal plan with fixed lunch",
        "description": "Generates a meal plan with a fixed lunch ID.",
        "parameters": [
          {
            "name": "mealPlanFixedLunchDetails",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dietitian_ID": {
                  "type": "string"
                },
                "protein_goal": {
                  "type": "number"
                },
                "carbs_goal": {
                  "type": "number"
                },
                "fat_goal": {
                  "type": "number"
                },
                "nbr_days": {
                  "type": "integer"
                },
                "fixed_lunch_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/GenMealPrep"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/MealPrep/insertMealPlanItem": {
      "post": {
        "tags": ["Meal Plan Item API"],
        "summary": "Adds a Meal Plan Item to the DB",
        "description": "Adds a Meal Plan Item to the DB",
        "parameters": [
          {
            "name": "mealPlanDetails",
            "in": "body",
            "required": true,
            "schema": {
                "$ref": "#/definitions/MealPrepItem"
              }
          }
        ],
        "responses": {
          "200": {
            "description": "meal prep item added successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }, 
    "/MealPrep/insertBulkMPItems": {
      "post": {
        "tags": ["Meal Plan Item API"],
        "summary": "Adds several Meal Plan Items to the DB in one call",
        "description": "Adds several Meal Plan Items to the DB in one call",
        "parameters": [
          {
            "name": "insert_bulk_MP_Items",
            "in": "body",
            "required": true,
            "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/MealPrepItem"
                }
              }
          }
        ],
        "responses": {
          "200": {
            "description": "meal prep items added successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }, 
    "/MealPrep/getCombination": {
      "get": {
        "tags": ["Meal Plan Item API"],
        "summary": "gets a specific saved combination based on the patient, diet, and combination ID",
        "description": "gets a specific saved combination based on the patient, diet, and combination ID",
        "parameters": [
          {
            "name": "diet_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the diet"
          },
          {
            "name": "patient_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the patient"
          },
          {
            "name": "combination_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the combination"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Combination"
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/diet/getDietCombinations": {
      "get": {
        "tags": ["Meal Plan Item API"],
        "summary": "gets a all saved  combinations based on the patient and diet",
        "description": "gets a all saved  combinations based on the patient and diet",
        "parameters": [
          {
            "name": "diet_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the diet"
          },
          {
            "name": "patient_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the patient"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
                "type": "array",
                "items": {
                "$ref": "#/definitions/Combination"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/Recipee/getRecipee": {
      "get": {
        "tags": ["Recipee API"],
        "summary": "Fetch Recipee details",
        "description": "Fetch details of a Recipee by its ID.",
        "parameters": [
          {
            "name": "recipee_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the recipee"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Recipee"
            }
          },
          "400": {
            "description": "recipee ID is required or other bad request."
          },
          "404": {
            "description": "recipee not found."
          }
        }
      }
    },
   "/Ingredient/details": {
      "get": {
        "tags": ["Ingredient API"],
        "summary": "Fetch ingredient details",
        "description": "Fetch details of a specific ingredient by its ID.",
        "parameters": [
          {
            "name": "ingredient_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the ingredient"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          },
          "400": {
            "description": "Ingredient ID is required or other bad request."
          },
          "404": {
            "description": "Ingredient not found."
          }
        }
      }
    },
    "/Ingredient/addIngredient": {
      "post": {
        "tags": ["Ingredient API"],
        "summary": "Add a new ingredient",
        "description": "Add a new ingredient to the database.",
        "parameters": [
          {
            "name": "ingredient_data",
            "in": "body",
            "required": true,
            "description": "Data of the ingredient to be added.",
            "schema": {
              "$ref": "#/definitions/CreateIngredient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingredient added successfully."
          },
          "400": {
            "description": "Bad request."
          }
        }
      }
    },
    "/Ingredient/updateIngredient": {
      "put": {
        "tags": ["Ingredient API"],
        "summary": "Update an ingredient",
        "description": "Update details of a specific ingredient by its ID.",
        "parameters": [
          {
            "name": "ingredient_data",
            "in": "body",
            "required": true,
            "description": "Updated data of the ingredient.",
            "schema": {
              "$ref": "#/definitions/Ingredient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingredient updated successfully."
          },
          "400": {
            "description": "Ingredient ID is required or other bad request."
          },
          "404": {
            "description": "Ingredient not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    }, 
   "/Diet/createDiet": {
      "post": {
        "tags": ["Diet API"],
        "summary": "Create a new diet",
        "description": "Add a new diet record for a patient.",
        "parameters": [
          {
            "name": "diet_data",
            "in": "body",
            "required": true,
            "description": "Data for the diet to be created.",
            "schema": {
              "$ref": "#/definitions/CreateDiet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diet added successfully with ID: {diet_id}"
          },
          "400": {
            "description": "Bad request."
          }
        }
      }
    },
    "/Diet/udpateDiet": {
      "put": {
        "tags": ["Diet API"],
        "summary": "Update an existing diet",
        "description": "Update a diet record for a patient using the diet ID.",
        "parameters": [
          {
            "name": "diet_data",
            "in": "body",
            "required": true,
            "description": "Updated data for the diet.",
            "schema": {
              "$ref": "#/definitions/Diet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diet updated successfully with ID: {diet_id}."
          },
          "400": {
            "description": "Bad request or Diet ID is required."
          },
          "404": {
            "description": "Diet not found."
          }
        }
      }
    },
    "/Diet/deleteDiet": {
      "delete": {
        "tags": ["Diet API"],
        "summary": "Delete a diet",
        "description": "Remove a diet record using the diet ID.",
        "parameters": [
          {
            "name": "diet_id",
            "in": "body",
            "required": true,
            "description": "ID of the diet to be deleted.",
            "schema": {
              "type": "object",
              "properties": {
                "diet_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diet deleted successfully with ID: {diet_id}."
          },
          "400": {
            "description": "Diet ID is required or bad request."
          },
          "404": {
            "description": "Diet not found."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Dietitian": {
      "type": "object",
      "properties": {
        "dietitian_id": {
          "type": "string",
          "description": "ID of the Dietitian."
        }, 
        "first_name": {
          "type": "string",
          "description": "First name of the Dietitian."
        },
        "family_name": {
          "type": "string",
          "description": "Last name of the Dietitian."
        },
        "date_of_birth": {
          "type": "string",
          "format": "date",
          "description": "Date of birth of the Dietitian."
        },
        "phone_number": {
          "type": "string",
          "description": "Phone number of the Dietitian."
        },
        "email": {
          "type": "string",
          "description": "Email address of the Dietitian."
        }
             }
    },
...
  "GenMealPrep":{
    "type": "object",
    "properties": {
      "best_combinations": {
        "type": "array",
        "description": "Array of Breakfast - Lunch - Dinner Combinations.",
        "items":{
                "$ref": "#/definitions/Combination"
              }
            }
          }
        },
...
      "apiResponse":{
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "success or error"
          },
          "message": {
            "type": "string",
            "description": "error message"
          }
          
      }
    }
      }
}